Automatically generate a schema file by default
Since we're deprecating `infer_schema!`, we want to generate this file
by default. I think we should eventually change our behavior on tables
with no primary key to warn and skip instead of error, but I want to do
a bigger restructuring once we can get rid of `infer_schema!`.