Add a helper trait for writing `ToSql` impls for PG tuples
I had meant to do this when adding support for `Record`, but didn't get
around to it. However, someone in gitter needed this today, so now is as
good a time as any.

We avoid implementing `ToSql` directly, because we don't want
`Bound<Record<(Integer, Text)>, (i32, String)>` to become valid.