Fix the handling of unsigned values on MySQL
We were sending the right bytes, but we weren't properly setting the
`is_unsigned` flag on the bind parameters, resulting in it always being
sent as signed. This results in a client error about overflow when
deserializing large values, and either a server error or incorrect
results when serializing them.

The fact that the tests for this all used `-1` should have been a huge
red flag, but it slipped through.

Unfortunately, there's no way for us to reasonably fix this without a
breaking change or crazy hacks (we could do something like write an
additional byte at the end if it's unsigned, but that's probably not
worth it to avoid a breaking change). For that reason, we will need to
make the next version of Diesel 2.0. We shouldn't go crazy with this
change, but this would also allow us to make #1457 happen as well.

The other alternative would be to introduce a `Mysql2` backend, and
deprecate the current one.