Update to nightly-2018-07-17
I would have preferred to do this in smaller pieces, but unfortunately
we had to do this all in one go. Clippy is no longer publishing to
crates.io, so we have to switch to using `cargo clippy`. There are also
a few bugs in recent versions that we have to work around.

The version of proc-macro that we were using no longer compiles on the
most recent nightlies, so we had to bump that as well (and therefore
also bump our version of `syn`). The biggest change here is that
`syn::Ident` is now `proc_macro2::Ident`, which does not implement
`From<&str>`, `AsRef<str>`, or `Copy`.

Finally, our UI tests regressed significantly, as Rust is no longer
giving us any useable span information for any piece of attributes in
derive macros. There's nothing we can do to work around this, we'll just
have to wait for Rust to fix it on their end.