Remove some UB from our statement cache
From https://doc.rust-lang.org/beta/reference/behavior-considered-undefined.html

- Invalid values in primitive types, even in private fields and locals:
  - Dangling or null references and boxes.

This means that our "create an `&mut T` pointing to arbitrary memory"
strategy is UB, even if we never dereference it. Instead of trying to
lift the error out of `RefMut::map`, this new version instead calls the
function outside of `RefMut::map`, allowing us to just do normal error
handling. We then turn the result into a raw pointer to eliminate its
lifetime, so we can stick it back into a `RefMut` later.

This does create a pointer which alises a live `&mut` binding, but as
far as I can tell, this is only UB if we access both -- which we don't.