Allow `now` to coerce to `Nullable<Timestamp>`
It is able to coerce to `Nullable<Timestamptz>`, so it makes no sense
for it not to coerce to `Nullable<Timestamp>`.

I'd love to use `expression::Nullable<now>` both in the test and as the
return type of the coercion (or remove the nullable coercion entirely).
However, `expression::Nullable` can only be used in select clauses with
left outer joins for complicated coherence reasons