Fix compatibility with MySQL 8.0
MySQL's client library puts the responsibility on us to clear the
network buffers before continuing. This is problematic in cases where we
execute a query that returns results (and thus has additional frames
that need to be flushed), but in a context where we're ignoring them.

The old implementation used `mysql_next_result`, which advances to the
next result in the query, and returns whether we just advanced to the
final result or not. Because it returns false on the last result, and
not if there are no more results to advance to, we had to call
`mysql_store_result` a final time after the last call to
`mysql_next_result`.

The problem case is when a query only has a single result. In MySQL 6
and earlier, the second call to `mysql_store_result` would return a null
pointer. In MySQL 8, it will return an error instead. To fix this, I've
separated out the "more results", which will give us the behavior we
wish `mysql_next_result` had.

I'm honestly not sure why migrations weren't affected by this, as I'd
expect to get the same behavior in a migration that contained only a
single statement with no results, but apparently this only occurs if
there is a statement that does have results.

I have not updated the version of MySQL on Travis, because I want a CI
run to make sure I didn't break things on 6.x. I will update it to point
at 8.0 before merging.

This change does not require additional tests, as we already have
regression tests in place for this case.

This fix justifies a patch release backported to 1.3.

Fixes #1826.