Add `--verify-schema` to CLI, for use in CI and production
While the config file and auto-updating schema file is a nice quality of
life change, it's too easy to accidentally have the contents of
src/schema.rs not reflect reality, but have CI pass because the file
gets changed when the test setup runs.

This adds a new flag, `--verify-schema` which causes any command that
would update the schema file specified in your config to error if the
contents changed as a result of that command. It's expected that this
command would always be run in CI and production deploys.