Update our issue policy to make the issue tracker more manageable
As per the discussions we've been having in gitter, this is a change
that is meant to make the issue tracker more manageable, and more
useful at surfacing features we should work on next.

The goal is to have every open issue be something immediately
actionable, meaning it is either a bug, or a feature we are actively
planning to work on in the near future. Doing this effectively turns the
issue tracker into a todo list, where any open issue is something we can
actually work on today. It also gives better visibility into our
roadmap, so folks can see what is coming in upcoming versions, vs things
that sound reasonable for some point in the future.

The other problem with using the issue tracker for feature requests is
how binary open/closed is. We have to immediately decide whether a
feature is reasonable for some point in the future, even before
discussion around it happens. If we close the issue, that likely kills
any discussion, but leaving it open implies that it's a feature we'd
accept a PR for.

Feature requests are something we still want to accept, but we will only
accept them on our discourse going forward. This moves the discussion to
a platform more designed for those sort of open ended discussions. It
also effectively creates an informal RFC process, where asking someone
to open an issue with their feature requests is effectively an accepted
RFC, while leaving a thread open without explicitly rejecting the
feature is effectively a postponed RFC.

After this is merged I will go through all open issues and enforce this
policy.