Remove support for libsqlite3-sys 0.10.0
Similar to #1900 libsqlite3-sys 0.10.0 is not selected by cargo because
it misses the `min_sqlite_version_3_7_16`. Similar to the situation
there we need to continue to support the old versions (Not because it's
part of a public api in this case, but because it is only possible to
link one version of a native library).