diesel_cli: Change from Error::cause() to Error::source()
As of 1.30.0 there is Error::source() to replace Error::cause().
The former adds a 'static lifetime requirement.  As of 1.33
the latter is deprecated which means the nightly test runs fail.

Since the min-version supported by Diesel is now 1.31.0 we can
switch to Error::source() and thereby avoid the deprecation warnings.

Signed-off-by: Daniel Silverstone <dsilvers@digital-scurf.org>