Introduce a derive for `NonAggregate`
This introduces a derive for the `NonAggregate` trait, which we're
planning on removing as laid out in
https://discourse.diesel.rs/t/proposed-change-replace-nonaggregate-with-something-less-broken-that-can-support-group-by/18.
The purpose of this derive is mostly to reduce the churn caused by the
changes laid out there in the most common case. Because we don't yet
have a PR opened for that change, and it's not clear that is actually
the direction we will end up going, I have written this as a standalone
change that can be shipped in terms of documentation and CHANGELOG
entries.

As part of this, I reviewed every impl of `NonAggregate`. A *lot* of
these impls were missing the bounds required to make them sound. That's
been fixed for most of them by using the new derive. A few others that
can't use the derive have been fixed manually. In a lot of cases there
was also a useless `Self: Expression` bound which is no longer required
and has been removed.