Cleanup the Backend trait
The `ByteOrder` associated type is documented as only be used if the
`RawValue` is `[u8]` but that's was never true.
We've now a trait (`BinaryRawValue`) that indicates if a backend uses
raw bytes as raw values representation and therefore has a use case
for the `ByteOrder` associated type, so move that type
there. Additionally fix those cases where the type was used
"accidentally" even if that results in some code duplication.