Make generated schema the same on all platforms
Resolves #2202

It appears that the following query does not behave the same way on all platforms:
```sql
SELECT table_name, table_schema from information_schema.tables
WHERE table_schema = 'your schema' AND table_type LIKE 'BASE TABLE' ORDER BY table_name;
```

The order changes depending on the platform and the version of postgres (more details [in the PostgreSQL bug mailing list](https://www.postgresql.org/message-id/11505.1571936496%40sss.pgh.pa.us)).

This query is used when generating the `schema.rs` in the Diesel CLI.
The fact the order changes depending on the platform creates issues when applying, say, an unified-diff patch to the schema afterwards.

This patch sorts the table names on the rust side so that the order of the tables will not depend on the collation of the `information_schema` table.

PostgreSQL >= 12 has *by default* on all platforms a collation of the `information_schema` table that results in the same generated `schema.rs`s as this patch enforces on all versions/platforms/configurations of Postgres, so this is a *one-shot* controlled somewhat-breaking change that is otherwise already happening *gradually* on all platforms.