Migrate the main crate to the 2018 edition.
Note that this does not make any idiom adjustments, nor does it make the
crate usable without `#[macro_use] extern crate diesel;`. This is the
result of running `cargo fix --edition`, fixing the broken code it
generated, and then manually fixing the doctests.

As part of this I noticed a few doctests that expected the `extras`
feature to be able to run. I fixed those, and also realized that the
`Macaddr` type's impls are hidden behind the `network-address` feature
for no real reason.