Update our macros to work in the 2018 style
This updates our crate to work without `#[macro_use] extern crate
diesel` at the crate root. I've audited all uses of `#[macro_export]`.
If it was obvious at a glance that inner macros are local, I've added
`local_inner_macros`. Otherwise I changed it to use `$crate::`. Most of
the non-public macros can probably have the `__diesel_` prefix removed,
but I've left it on to keep things reasonable for 2015 edition users and
to minimize the diff.

We actually have a surprisingly small number of non-derive macros that
are public API. It's pretty much limited to our schema macros and
operator generation. The schema macros are exported from the prelude,
the operator generation macros are not as they had their prefix removed
in 2.0 with the expectation they are always invoked with `diesel::`.

I've removed all instances of `extern crate diesel` from the doctests.
Note that this does not change anything outside the main crate. In
particular, `diesel_cli` will still generate a `schema.rs` that assumes
`#[macro_use] extern crate diesel;`. It will need to import the prelude
at the top of that file. This will be addressed once that crate is moved
to the 2018 edition.