Refactor `mysql::StatementMetadata` to remove foot guns
This type was highly unsafe to be used. It was possible to easily
return dangling pointers to field names for example. This change tries
to wrap the unsafety inside the `mysql::connection::stmt::metadata`
module by provide a wrapper type to interact with fields returned by
`StatementMetadata` and by removing as much of the `'static` hacks
used there.