Cherry pick "Fix transactions not closing properly when failing on commit" from master c7546f802a86613bc4cdd914a3fa24d4a165cce5
Connection could be made available again with a still-open transaction due to transaction commit fail (typically `SerializationFailure` or `ReadOnlyTransaction`).

This eventually triggered "AlreadyInTransaction: Cannot perform this operation while a transaction is open" errors in unrelated places.

This commit fixes this issue by catching `SerializationFailure` on commit and attempting a rollback. (`ReadOnlyTransaction` was however not backported.)