Allow to return non nullable values from a custom aggregate function
If a custom aggregate function returns a default value if an empty set
is passed as argument there is no need to return a nullable value
here. This change allows to return non nullable value directly from
the implementation of `SqliteAggregateFunction::finalize`. It is clear
from the presence of `Option<Self>` as argument that there may not be
any instance of `SqliteAggregateFunction` there and some default/none
handling is required.