Add support for dynamic select clauses to diesel_dynamic_schema
This commit adds support for constructing dynamic select clauses and
receiving values into a dynamically sized container. I've opted for
requiring a user provided type as value type in the dynamically sized
`DynamicRow<I>` container, because that type needs to handle all
possible values that can appear in the loaded row. The requirements for
such type are normally highly dependent on the actual use case,
therefore we do not provide a "covers all cases" container type yet. The
implementation leave open the possibility to add such a container type
by just setting a default type for the generic paramater there.

See the added test cases and documentation for examples how to use this feature.