Fix #2140
Include parenthesis around every infix operation provided by one of the
`ExpressionMethod` traits.

It is not possible to include this directly into the `infix_operator!`
macro due several SQL constructs:

* `a LIKE b` can have optional `ESCAPE` clauses (same for `NOT LIKE`,
`ILIKE` eg). Unconditionally putting a parenthesis around `(a LIKE b)`
is no valid SQL
* `a BETWEEN l AND u` is implemented a set of two infix operators,
`BETWEEN` and `AND`, so putting the parenthesis directly into the
operator macro would lead to `x BETWEEN (l AND u)` which is not valid
SQL

Instead I've opted for adding explicit parenthesis in the corresponding
`ExpressionMethod` (and similar traits` implementations. As part of this
change I've changed all the return types there to use some type from
`diesel::dsl`, so that potential users have an easier time to figure out
what types to use if they have to write some where clause or something
like that.

While changing the return types to `diesel::dsl` types I noticed  and
fixed the following issues:

* Not all methods had an equivalent type in `diesel::dsl`, sometimes
because it was missing, sometimes because the postgres specific types
weren't correctly exported there. Both is fixed now
* Some types were not anymore compatible with the actual methods. A
notable example is `dsl::And` and `BoolExpressionMethods::and` where the
later returned a potentially nullable type depending on the input types,
while the first was only valid for not nullable right sides. Fixing that
was a bit more complicated:
    * I've opted for making the return type of `BoolExpressionMethods::and`
      (and `or`) unconditonally nullable. This prevents having an explicit
     `ST` parameter on `dsl::And`
    * This in turn requires that we accept a non nullable expression in
      places where a nullable expression is expected. Technically this
      was already possible, but would require explicitly calling
      `.nullable()` on affected expressions. Given that `.and()` and
      `.or()` are really common, that seemed not to be a good solution.
      As result I've changed the `AsExpression` impl for `T: Expression`
      in such a way that for any `T` with `T::SqlType: SqlType<IsNull =
      is_nullable::NotNull>` there is now an impl
      `AsExpression<Nullable<T::SqlType>>` so that such expressions are
      also accepted in nullable contexts.