Fix handling of numeric types for mysql
Newer libmariadb versions do perform not that much conversion logic on
requested binds, so does for example `select 1/2` result in `0.5`. As
result of a prior change we already have most of the logic to do the
correct conversion in diesel in place.
As part of this change we request change now the requested type to the
actual field type as returned by libmysqlclient/libmariadb and do the
conversion ourself as long as we are sure that the types are compatible.
Additionally the conversion code now handles cases where we receive a
number that would not be representable using the target type. For such
cases a `Numeric overflow/underflow` error is returned