Add a `Selectable` trait
This commit adds a `Selectable` trait, a corresponding derive and some
methods that allow to construct the select / returning clause based on a
type implementing this trait before sending the corresponding query to
the database.

This trait is designed to be used in combination with `Queryable` which
represents a result of a query. Ultimativly my hope here is that this
combination drastically reduces the number of type missmatches that
occure because of missmatches between query and struct implementing `Queryable`.