Return a boxed iterator from `RunQueryDsl::load_iter`
This makes it much more understandable what's the type of the return
type, while having a minimal performance cost.
I've opted into applying this change only for this method and not for
`LoadDsl::internal_load` and `Connection::load` as both are not really
user facing. (I would count them as advanced methods most users never see
and never use directly)