Move the PgConnection and MysqlConnection behind a distinct feature flag
This allows third party backends to implement custom Connections without
depending on libpq/libmysqlclient (through PgConnection and
MysqlConnection) while reusing the existing backend implementations for
postgresql and mysql

Additionally this change makes a few details required for the
implementation of such backends public. As a test case I've implemnted a
MysqlConnection based on the mysql-simple crate and a postgres
connection based on the rust-postgres crate as proof of concept. (I plan
to submitt those connection implementations as PR's to the corresponding
crates after the 2.0 release.)

The same split is not done for the sqlite backend as the connection
implemenation must be based on libsqlite there for all cases.