Fix #3016
This demonstrate a potential fix for #3016 by using the async libpq
interface + busy polling. This allows us to abort polling as soon as we
reach the timeout.
I see multiple issues with this approach:
* Add's new methods to `Connection` (`get_timeout`/`set_timeout`). This
  could likely be avoided by exposing them directly on `PgConnection`.
On the other hand that would, at least for 1.4.x imply that we cannot
use those timeout only for the `is_valid` function. It would be all or
nothing then.
* This uses busy polling, which could be quite inefficient. I'm not sure
  what better posibilities there are without pulling in
tokio/mio/async_std.