Revert to use mariadb 10.5 on the macos CI for now
I do not have any idea why 10.6 fails to run this specific query.
Everything seems to be right. In contrast to what the error message
suggests the problem seems to be on the server side as just changing the
client lib does not solve the problem. At least for me that make it more
plausible that we just hit an bug in their server implemntation.