Improve the index implementation for Jsonb::remove
This removes the need to explicitly specify the sql type of the key by
introducing a new trait for this. The downside of the new approach is
that custom types that map to `Text`/`Integer`/`Array<Text>` won't be
supported there automatically. They would need a custom impl of that
trait. I feel that's fine because the trait covers hopefully the
majority of what user would try in this position.