Inline `FromSqlRow::build_from_row` as this function seems to be really hot
This seems to give up to 5% reduction in instruction count for the
`diesel/trivial_query/1000` benchmarks. Similar improvements are
observed for the boxed query variants.
This change seems to be neutral for the `queryable_by_name` variant.
For the insert/1 benchmark a slight regression in instruction count can
be obsverved.