Add support for slices of refs with BelongingToDsl
To support this usage we need to implement Identifiable for references
of references as well, thus we add a blanket impl here.
Furthermore, BelongsTo needs to be implemented for &Parent.
As blanked implementation for this is only possible with workarounds to avoid
overlapping implentations, we instead add an additional implementation to
the derive macro.

Co-authored-by: Konstantin Baltruschat <k.baltruschat@heinlein-video.de>