Add support for slices of refs with BelongingToDsl
See #3138 for discussions.

We cannot add a blanked impl like
```rust
impl<'a, T> Identifiable for &'a &'a T
where
    &'a T: Identifiable
{
    // …
}
```

here as this results in type inference issues with `.belonging_to` and
`.grouped_by`. Rustc tries to infinitely recurse on `&&&&&…&&& T as
Identifiable` then. As of this I've opted into just generating concrete
impls via `#[derive(Identifiable)]` as this works around this type
inference issues.

Also we cannot use a wild card impl for the added `BelongsTo<&Parent>`
case as this causes a conflicting impl error with our
`BelongsTo` impls for tuples. So again we use concrete impls here.