This commit fixes a potential race condition in our r2d2 tests
This was triggered by the fact that the pool tries to open a new
connection after the old one was closed. As this happens in a background
thread this happens concurrently to our test code. That means if
establishing a new connection is fast that will increase the
acquire_count before we checked it. We observed this happening for
sqlite. I've opted into just not checking the acquire count there as
this is not really relevant anyway at that points.