Remove unused lifetime from derived impls
This change removes a generated lifetime from the derived
implementations of Insertable and AsChangeset for owned structs. These
lifetimes should only be added to the implementation over references.

On newer versions of Clippy (with the release of Rust 1.62.0, as of this
commit), this results in a lint error: extra_unused_lifetimes.

See
https://rust-lang.github.io/rust-clippy/master/index.html#extra_unused_lifetimes
for more information.

Signed-off-by: Peter Schwarz <pschwarz@bitwise.io>