CLI: Allow running without a `down.sql` migration
Some people don't like to write down migrations, as depending on your workflow, they take a lot of time to write, are *never* used (esp. when you have CD pipelines), and consequently are typically largely untested.

We support this for those (like us) who would rather write a new `up` migration to revert unwanted changes through CI/CD if necessary.

(NB: The original issue is that since we wouldn't fill down.sql, we would have to go edit that file to add a newline for formatting everytime we create a migration, and then leave it there even though it doesn't mean anything.)