Allow custom select expressions in `Selectable`
It's is a recurring use-case that 90% of the fields of a `Selectable` struct could be selected directly, but just one of them needs a bit more complex expression.
For this use case, this commit adds support for specifying fully custom select expression at the field level when deriving `Selectable`, by specifying the `select_expression` and `select_expression_type` attributes.