add time dependency impls mirroring chrono's for sqlite
WIP: debugging ToSql

This is just getting more confusing: it looks like it's encoding to
precisely the representation we want, but sqlite still thinks that
these values are unequal.

fix sqlite time encoding test

Turns out that Sqlite doesn't expect / desire that we write the
subseconds if they are equal to 0, apparently. This is a confusing
and frustrating issue, but at least we figured it out! Just use a
format which doesn't expect to write subseconds if the subseconds
should be zero.

The rest of the tests are likely to be fixed with analogous measures.

fix sqlite time tests

Looks like Sqlite is very particular about how many subseconds are
recorded when saving timestamp information, but we can conform to
its expectations now that we know what they are.