fix sqlite time encoding test
Turns out that Sqlite doesn't expect / desire that we write the
subseconds if they are equal to 0, apparently. This is a confusing
and frustrating issue, but at least we figured it out! Just use a
format which doesn't expect to write subseconds if the subseconds
should be zero.

The rest of the tests are likely to be fixed with analogous measures.