Fix #3252
This commit fixes an issue where test transactions caused connection to
be marked as broken when returned to a r2d2 connection pool. It is a
quite common pattern to create a connection pool with 1 connection, and
call `begin_test_transaction` on the connection before putting it into
the pool. This allows users to easily test applications using connection
pools without modifying an existing schema. The recent changes to our
transaction manager implementation + our r2d2 implementation broke this
pattern.

This commit introduces an additional flag inside of the transaction
manager state, which tracks whether `begin_test_transaction` was called
or not. This flag is later used to decide whether a connection should be
considered broken or not.

In addition this commit centralises the `is_broken`  implementation as
part of the transaction manager itself + marks
`transaction_manager_status_mut` as part of the unstable public
API (this method is only useful for third party connection or
transactionmanager implementations)