diesel_cli: handle file: URLs in sqlite db setup
Fixes #3301.

The previous code in create_database_if_needed assumed the string
argument 'database_url' was a path in the sqlite case. This needs to be
supported since it's the only way to use a relative path with
diesel_cli, but it led to surprising behavior when passing a file:///
URL -- when checking for the existence of the DB, diesel_cli would look
for a relative path file:/foo/bar where it should've looked for an
absolute path /foo/bar.