Make the `ValidGrouping` impl for `AliasedField` more generic
This commits makes the `ValidGrouping` impl for `AliasedField` more
generic so that diesel now also accepts queries which group by a primary
key and select any other field of the alias. This change is backward
compatible even with the changed trait impl as any type that satisfied
the old impl also satisfies the new impl for `C1` == `C2`.

This fixes #3319