Verify that group_by expression is valid for the given table
This commit fixes a bug where diesel would accept group by expressions
that contain columns of tables not even appearing in the current from
clause. The provided compile test contains some examples that would have
been accepted before this commit (not all of them where accepted before,
only those with `.select(…).group_by(…)`).

This is technically a breaking change as this rejects previously
accepted queries. However I would argue that this is at the same time a
critical bugfix as this turns invalid queries in a compiler error.