Fix #3330
This commits fixes an issue with our dsl generating an invalid query for
`UPSERT` statements with `WHERE` clauses. As described in
https://www.postgresql.org/docs/current/sql-insert.html `WHERE` is only
valid in combination with `DO UPDATE` here. In detail this commit fixes
the following issues:

* `WHERE` appears now in the correct place in the generated query (after
`ON UPDATE SET …` instead of before this clause
* We now reject queries that use `filter_target` but `DO NOTHING` as
action. Those are not valid as outlined in the postgresql documentation.

In addition a test case for both issues is added.