Add some additional checks to `#[derive(Selectable)]` to ensure field compatibility
We generate an additional function with trait bounds for each fields to
check whether a certain rust field type is compatible with the
underlying database type. This greatly improves error messages for
potential type mismatches, as the user now gets an error message that
points to the field causing the problem + which has a pretty clear
meaning (stating that `FromSql<…>` is not implemented).

Unfortunaly that's not a perfect solution as we need to check
against **concrete** backend implementation. This PR solves this by
hard-coding all three out of the box supported backends (behind the
corresponding feature flags). That works for cases where users want to
have compatibility with all enabled backends, but there are certainly
use-cases where this is not the case. We might need some opt-in/opt-out
system here to prevent this from being a breaking change or making it
impossible to support valid use-cases. On the one hand we might want to
have this as default as the improved error messages is something which
is really helpful for new users, on the other hand I cannot see how that
would be possible without breaking changes.

Another down-side of this approach is that it generates basically
duplicated error messages when more than one backend is enabled.