Add a `#[derive(MultiConnection)]` proc-macro
This proc macro implements `Connection` and related traits for an enum
of connections for different database backends. This allows to easily
support multiple database systems at the same time. I've chosen to
implement this feature as a derive, instead of providing just one
`AnyConnection` enum with all diesel connection types, to allow third
party connection implementations to be used there as well.