Fix #3470
This commit tries to address an issue where we do not issue rollback to savepoint
commands due to libpq reporting a transaction as broken. It turns out
that transactions are sometimes recoverable by just rolling back to a
previous savepoint instead of rolling back the outer transaction.

We address this issue by always executing the `ROLLBACK TO SAVEPOINT`
command even if a connection is required as needing a top level
rollback. This was previously skipped. We do ignore any errors from that
attempt if the requires top level rollback flag is set. If the `ROLLBACK
TO SAVEPOINT` command succeeds, we consider the transaction as recovered
and disable the requires top level rollback flag.