Fix #3558
This commit fixes an issue where out of bound text/binary binds on `sql_query` using
the sqlite backend causes a panic in an internal drop impl. It was
caused that we pushed to the list of released binds before we actually
checked whether that bind was possible or not. This caused us trying to
bind to the same invalid bind again in the drop impl, while being on
the normal error path. It's fixed by just moving the relevant push after
the actual bind. In addition a regression test for this problem is added.