Fix type resolution in rust-analyzer
It turns out that we relay on a type resolution hack for our derives
that only work with rust 2015. That's not a problem with rustc, as it
uses whatever edition a crate is using. So it will use rust-2015 for
name resolution in the expansion of our proc macros, due to the fact
that `diesel_derives` is a edition = 2015 crate. The problem is that
rust-analyzer does not understand this, as they use basically the name
resolution rules from the 2018 edition. (To be clear that's something
that's not "standard" conform there, but it does not seem to be a
priority to fix that).
As this affects our core type system (== all the SQL types, …), that has
pretty serve effects on what users see in their IDE.

Fortunately the workaround is simple as we just need to use an
alternative way to provide an diesel item in diesel itself to make our
macros work there. It seems like both rustc and rust-analyzer agree on
the `extern crate self as diesel;` solution.

After this basically all known cases of rust-analyzer issues are gone
away (at least as far as I tested).