Fix #3579: Unexpected end of row when setup diesel
We use `PRAGMA table_info` for older versions of sqlite. This returns
fewer columns that `PRAGMA table_xinfo` which we use in newer versions
of sqlite, making the two code paths incompatbile with the `Queryable`
implementations. We instead now use `sql_query` with an implementation
of `QueryableByName` to return default values for the case of `PRAGMA table_info`.