Introduce support for connection instrumentation
This commit adds functionality that allows to add a relatively fine
instrumentation to our connection types by providing an essentially call
back based pattern for instrumentation. The implemented setup calls the
provided instrumentation type with different events. This allows the
instrumentation to decide on it's own which events are important and
which are unimportant. It also enables to skip most of the work (like
constructing the sql of an inspected query) if the event is not handled
as we just pass down an opaque wrapper that can be evaluated by the
instrumentation implementation.

This commit includes:

* A default instrumentation implementation that does nothing
* A global way to set the instrumentation implementation used by new
connections
* A connection specific setter to change the instrumentation
implementation for a specific connection
* A wild card instrumentation implementation for closures that accept
the event type

This commit does not include any "advanced" instrumentation
implementations (based on `log` or `tracing`, etc). The idea is
that these live in their own crates as it is might depend on the actual
use case how the different events should be handled.

The implementation of `InstrumentationEvent` is decoupled form specific
backend types to allow reusing the same instrumentation for different
connection types. The definition of `Instrumentation` does not depend on
any connection specific stuff so that it is possible to use the same
implementation for `diesel-async` as well.