Fix #3790
This commit fixes an issue in `#[derive(Selectable)]` that caused
generating invalid rust code for structs with lifetimes (and type
parameters). It improves the derive to handle type parameters and
lifetimes correctly. In addition it also improves the derive to generate
an error if we find a reference, as plain references are not supported
as load target by diesel.