Optimise the CI pipeline
This commit tries to reduce the time to get feedback whether or not a
build fails by only running all the tests on all platforms if basic
checks succeed. For that we first run the rustfmt + clippy check (that
builds basically all the code) and the `sqlite bundled` check (that runs
most of the tests on linux + sqlite).