Refactor error handling in diesel-cli
This commit refactors the error handling in diesel-cli to use a central
error type and to use errors wherever meaningful (in contrast to just
panicing). This hopefully improves the quality of the error messages
emitted by diesel-cli

In addition to the error handling changes this commit also includes some
log statements for:

* Complex functionality
* Any query executed

This hopefully simplifies debugging any potential issues in the CLI tool
more easily.