Fix several issues with `#[derive(MultiConnection)]`
This commit fixes several issues that occur with multiconnection
implementations:

* We now generate an implementation of
`Connection::begin_test_transaction` that just calls the inner
implementations as the inner connection implementations might
overwrite the default implementation

* We now mark queries as unsafe to cache if they are in marked as unsafe
to be cached for the specific backend.

* Another fix for binding null values as the assumption that the actual
types don't matter seems to be not correct. I've replaced the hack with
an distinct function on the inner bind collectors itself instead.

* Another fix for the the type metadata returned by `HasSqlType` as that
could result in calling the impl for the wrong backend if backends share
the same `MetadataLookup` type. The previous implementation did just use
the first backend that returned a concrete type metadata lookup value,
this patch just calls all possible backends and return all possible
values in a struct instead of an enum.

In addition this PR adds some doc comments to some locations that are
warned of while building with the
`i-implement-a-third-party-backend-and-opt-into-breaking-changes` flag