Bump pq-sys to 0.5
This commit increases the maximal supported version of pq-sys to include
support for 0.5. That release contains bundling support for libpq and
support for building with address sanitizers enabled. This PR adds the
relevant features to diesel_cli to allow using the bundling support + it
adds a CI job for also running tests with address sanitizers enabled.