Make more tests deterministic
This commit fixes a lot of randomly failing tests by adding `ORDER BY`
clauses to some queries. It also prevents concurrently creating an
`INDEX` for the same table (which results in a deadlock in postgres
sometimes).

(This was tested by running `diesel_tests` in a loop until 100
continuous runs passed)