Add implicit sqlite rowid columns to print-schema output
Whenever sqlite uses an implicit rowid it also need to be added as
column to the corresponding schema. This change fixes the implicit
primary key generator added in #3680 which was never tested as the test
added in this pull request never actually is run.

This change enables the test and enhances the schema printer to generate
the missing column if the primary key is an implicit rowid column.

Related: #2149