Add Postgres COPY FROM/TO support
This commit adds support for PostgreSQL `COPY` commands.

For `COPY FROM` we expose a variant that allows users to configure the
stream manually and write directly to the stream. We also support a
variant that takes essentially a `Vec<Insertable>` (or equivalent batch
insert containers) and uses the binary format to perform a streamed
batch insert.

For `COPY TO` we expose again a variant that allows the user to
configure the stream manually and read directly from the stream. We also
support loading the results directly via the binary protocol into a
iterator of `Queryable` structs.