diesel_derives: specify `derive_from_sql_row()` docs
This change addresses a terminology detail that may slightly mislead
some library users. The current doc comment states that this derive
implements `Queryable` for primitive types. "Primitive type" in that
context is supposed to refer to a rust type that corresponds to a
single SQL type in contrast to a whole database row.

However, the current term unfortunately collides with Rust's definition
of "primitive types" [1] so we phrase it in a slightly more elaborate
way to prevent confusion.

[1] https://doc.rust-lang.org/rust-by-example/primitives.html