Change the order of generic arguments for InsertValues
This commit changes to order of generic arguments for the `InsertValues`
trait. This allows us to generate a sound impl of that trait via
`#[derive(MultiConnection)]` that will be accepted by future rust
compiler versions as well. This works as the backend type is this case
local, while the table type is generic. By having first the backend type
and then the table type we first have the local type, which is allowed
according to the rust coherence rules.

This is no breaking change as `InsertValues` is not part of the public
API: https://docs.diesel.rs/2.1.x/diesel/index.html?search=insertvalues

This fixes #4035