Fix lifetimes with `#[auto_type]`
This commit fixes using `#[auto_type]` with functions with reference
arguments. We now check the return type if any lifetime is contained
there and if that's the case we require that it is a named lifetime from
one of the function inputs. This lifetime is then correctly used in the
generated return type.