Fix a potential regression with old mysqlclient-sys versions
This commit fixes a potential regression with old mysqlclient-sys
version that just do not have a used constant. We fallback to using a
workaround based on default values in this situation. Manual tests with
all versions have verified that this seems to work
I'm not 100% sure why this was not found earlier by our minimal version check CI.