Add some tests that QueryableByName compiles with multiple SQL types for one output type.
This tests that `QueryableByName` compiles, and outputs some data, for one example
where the types are pulled from a `table!` macro; and another example where the
values are specified as annotations in the struct.

It's difficult to find a type that has the required properties across all three
databases; these tests are compiled only for Postgres. To do something similar for
the other databases would probably require compiling with one of the time libraries.