Add a `xtask clippy` and `xtask tidy` subcommand
The clippy subcommand just runs clippy for all workspace crates for a
specific backend (or all backends) with the right set of feature flags.

The `tidy` subcommand runs `cargo fmt --check`, `typos` and clippy for
all backends to provide an easy precheck before pushing the code.

This commit also adjusts the CI to use the `cargo xtask tidy` command
instead and the contributor gide to point to the new commends.