Fix a potential null pointer deref in the Sqlite backend
The documentation of `sqlite3_prepare_v3` states that:

> *ppStmt is left pointing to a compiled prepared statement that can be
executed using sqlite3_step(). If there is an error, *ppStmt is set to
NULL. If the input text contains no SQL (if the input is an empty string
or a comment) then *ppStmt is set to NULL. The calling procedure is
responsible for deleting the compiled SQL statement using
sqlite3_finalize() after it has finished with it. ppStmt may not be
NULL.

We already guard against the error case there before constructing the
`NonNull` pointer, we do not guard against empty statements yet. This
commit fixes that particular issue + adds a test to check if the fix
works as expected.

This fixes #4223