Support for window functions and other aggregate function expressions
See
[here](https://www.postgresql.org/docs/current/sql-expressions.html#SYNTAX-WINDOW-FUNCTIONS)
and
[here](https://www.postgresql.org/docs/current/sql-expressions.html#SYNTAX-AGGREGATES)
for the relevant postgresql documentation

This PR implements mainly the relevant DSL structure for this types. As
of today this supports most of the relevant API surface.

For now this PR only provides support for a single window function
(`count()`) but it brings in place all the necessary functionality to
support more window functions. It also updates any existing
`#[aggregate]` functions to mention the support of aggregate function
expressions.

Finally this PR deprecates the special `count_distinct` DSL function in
favour of the more general aggregate function expression implementation.