Refactor the statement cache abstraction
This commit refactors the statement cache abstraction so that it is now
possible to have a common statement cache implementation for diesel and
diesel-async. For this we need to abstract the
`StatementCache::cached_statement` method in such a way that it accepts
prepare functions that either return a `QueryResult<Statement>` or a
`impl Future<Output = QueryResult<Statement>>` depending on if the cache
is used in the sync or async implementation. 

For this we introduce a helper trait that allows us to apply certain
simple operations to both types. Diesel itself ships with an
implementation of that trait for `QueryResult<Statement>` while
diesel-async provides an implementation for the future variant.

As part of this change I also changed the caching strategy
implementations to not handle caching on their own as that would
require having the generic return type in the strategy trait as well 
which in turn would require having it on the statement cache as well,
which is not desirably. The strategy trait now only handles looking up
the cache key and returns something that indicates if the strategy
allows to cache the key (by exposing the hashmap `Entry` type) or not.