Make sqlite value function public
This commit turns the `read_*` functions of the `SqliteValue` type into
public functions, which makes it easier for third party crates to access
the underlying sqlite values in a performant way. This enables crates to
not clone for example the string, but rather reuse the string slice we got
from sqlite instead.

This also turns all these functions into methods that require a mutable
reference instead of a shared reference as technically each of those
methods might mutate the underlying value according to the sqlite
documentation.