Fix an issue that disallowed using `copy_from` with more than 12 column
This commit relaxes a trait constraint that restricted our copy from
implementation for `Insertable` types to 12 columns. That happened
because we used the `Default` impl for tuples from the rust standard
library to construct the column types internally. The standard library
only provides these impl for up to 12 tuple elements. 

The fix is to simply construct that type by calling `Default` for each
of the column types seperatly and then constructing the tuple from that.

I choose to not write a test for that as it doesn't change functionality
in any meaningful way other than just allowing larger tuples.