Refine stability rules for diesel release
This commit explicitly outlines that we might perform otherwise
considered breaking changes in minor releases if that's required by
build breaking changes from the rust project itself. A previous example
of such a change that nearly required us to make such a breaking change
in diesel is the change to default to `resolver = "2"` with the 2021
edition. I rather want to make sure that we don't end up in a similar 
situation in the future that requires us to issue a new major diesel version 
just to sidestep0breakage from their side.