fix: mark the r2d2 connection check query as safe to cache
currently diesel runs this query as a unnamed prepared statement
sending the parse and bind in different transactions. This breaks
when pgbouncer is used in between diesel and postgres.
https://github.com/pgbouncer/pgbouncer/issues/1288

Changing this to a named prepared statement does not affect
functionality in any way, and makes this work irrespective
of poolers.