Fix wtx loading_association bench
It turns out that this benchmark didn't insert the same amount of data
as the other benchmarks, so it run faster as it doesn't need to deal
with the same amount of data. This commit adds the relevant missing data
by inserting the 10k comments to the relevant table.

This commit also adds a few additional asserts to make sure that the
number of data is the same between diesel and wtx