Add `GroupedBy::try_grouped_by` method
`GroupedBy::try_grouped_by` is a proposed method for handling
grouping records, which may not have relationships to the provided
parents. It was proposed that this should be handled by collecting
these ungrouped records, and returning them as part of the result.

As a result, alongside adding the `GroupedBy::try_grouped_by` method,
This commit also adds the `TryGroupedByError` type, to contain
the results of error cases.

To avoid breaking changes, `GroupedBy::try_grouped_by` is also provided
with a default implementation, that calls the existing
`GroupedBy::grouped_by` method.