Add `TryGroupedByError` constructor
As `GroupedBy` is a trait, and can be implemented by external libraries,
these libraries may need to initialise an instance of `TryGroupedByError`
to implement `GroupedBy::try_grouped_by`.
As `TryGroupedByError` is marked with `#[non_exhaustive]`,
this requires a constructor function.

Additionally, this updates the `GroupedBy` implementation for `Iter`
to use `TryGroupedByError::new`, to minimise code-duplication,
and potential for future errors.