Add batch upsert support on sqlite
This commit adds support for batch upsert statements on SQLite if the
struct/insert type is marked as `#[diesel(treat_none_as_default_value =
false)]`. We don't provide an emulated variant for the other variant
yet, as results might differ if we just perform single statements in a
loop. Overall this brings the sqlite behaviour a tiny bit closer to the
postgres backend.