Add section on `GroupedBy` handling unrelated rows
All the current example of the usage of `GroupedBy::grouped_by`
are applied with the assumption it is being used with the
`BelongingToDsl::belonging_to` function, and consequently,
the child rows will be grouped using the same parent rows
they were queried using.

As rows can have multiple relationships,
and there is nothing preventing `GroupedBy` being used
with a different relationship to the one used to query a set
of child rows, this is not always true, and it is notable that
these child rows without a relationship in the provided slice,
will be dropped by the `GroupedBy::grouped_by` method.